requirements.txt
flask==2.3.2
pandas==2.0.3
PyPDF2==3.0.1
opencv-python==4.5.5.64
pytesseract==0.3.10
Pillow==10.4.0
